.history/src/predictor/requirements_20241202232000.txt
torch==1.7.1
torchvision==0.16.2
tensorflow==2.18.0
numpy==2.0.2
scipy==1.10.1
matplotlib==3.8.2
opencv-python==4.10.0.84
pandas==1.5.3
scikit-learn==1.2.2
pillow==10.0.1
tqdm==4.66.1
albumentations==1.3.1
